requirements.txt
bleach==3.1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
cycler==0.10.0
docutils==0.16
idna==2.9
importlib-metadata==1.6.0
jeepney==0.4.3
keyring==21.2.0
kiwisolver==1.2.0
matplot==0.1.9
matplotlib==3.2.1
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
pkginfo==1.5.0.1
pycparser==2.20
Pygments==2.6.1
pyloco==0.0.139
pyparsing==2.4.7
python-dateutil==2.8.1
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
SecretStorage==3.1.2
SimpleWebSocketServer==0.1.1
six==1.14.0
tqdm==4.45.0
twine==3.1.1
typing==3.7.4.1
urllib3==1.25.8
ushlex==0.99.1
webencodings==0.5.1
websocket-client==0.57.0
zipp==3.1.0
